Cleaned up organization of backend services, and implemented DB auto-initialization on startup.
