+++ visible/compose.yaml
@@ -15,6 +15,8 @@
     build:
       context: .
       dockerfile: visible.Server/Dockerfile
+    volumes:
+      - ./visible.Server/Data:/app/Data:ro
     ports:
       - 8080:8080
     links:
@@ -25,3 +27,4 @@
 secrets:
   postgres_user_password:
     file: secrets/postgres_user_password
+
