+++ visible/visible.Server/Configuration/SecretsConfigurationSource.cs
@@ -1,0 +1,12 @@
+namespace visible.Server.Configuration;
+
+/// <summary>
+/// Configuration source for Docker secrets 
+/// </summary>
+public class SecretsConfigurationSource : IConfigurationSource
+{
+    public IConfigurationProvider Build(IConfigurationBuilder builder)
+    {
+        return new SecretsConfigurationProvider();
+    }
+}
