+++ visible/visible.Server/appsettings.json
@@ -8,5 +8,6 @@
   "ConnectionStrings": {
     "PostgreDB": "Host=visible-db;Port=5432;Database=postgres;Username=postgres;Password="
   },
+  "DataPath": "/app/Data/",
   "AllowedHosts": "*"
 }
