+++ visible/visible.Services/Interfaces/IGigListingRepository.cs
@@ -1,4 +1,6 @@
-namespace visible.Services;
+using visible.Services.Models;
+
+namespace visible.Services.Interfaces;
 
 /// <summary>
 /// This class serves as the storage and access provider for Gig Listings.
